requirements.txt
chromadb==1.0.20
dashscope==1.24.1
fake_useragent==2.2.0
ffmpeg_python==0.2.0
Flask==3.1.2
Flask_JWT_Extended==4.7.1
flask_sqlalchemy==3.1.1
langchain==0.3.27
langchain_community==0.3.29
langchain_core==0.3.75
lmdb==1.7.3
moviepy==2.2.1
numpy==2.3.2
openai==1.106.1
opencv_contrib_python==4.10.0.84
opencv_python==4.10.0.84
opencv_python_headless==4.10.0.84
Pillow==11.3.0
pydantic==2.11.7
pydub==0.25.1
pytest==8.4.1
python-dotenv==1.1.1
Requests==2.32.5
scikit_learn==1.7.1
SQLAlchemy==2.0.43
tinydb==4.8.2
torch==2.3.0+cu121
torchaudio==2.3.0+cu121
tqdm==4.65.0
transformers==4.56.0
volcengine_python_sdk==4.0.10
Werkzeug==3.1.3
gunicorn==23.0.0
